Fix file permission issue
This fixes issue with permissions.  Because you run the install script as root, the file ownership of bin directory is root.  This leads to permission errors when you start a systemd process as a different user.  So I added a recursive chown of the AGENTDIR to set the owner as AGENTUSER.